+++ Assets/Scripts/UI/GamePage/YakuObject.cs
@@ -1,0 +1,18 @@
+using TMPro;
+using UnityEngine;
+
+namespace MCRGame.UI
+{
+    public class YakuObject : MonoBehaviour
+    {
+        [SerializeField] private TMP_Text yakuNameText;
+        [SerializeField] private TMP_Text scoreText;
+
+
+        public void SetYakuInfo(string yakuName, string score)
+        {
+            yakuNameText.text = yakuName;
+            scoreText.text = score;
+        }
+    }
+}
