--- Assets/Scripts/UI/YakuObject.cs
@@ -1,17 +1,0 @@
-using TMPro;
-using UnityEngine;
-using MCRGame.UI;
-using DG.Tweening;
-
-public class YakuObject : MonoBehaviour
-{
-    [SerializeField] private TMP_Text yakuNameText;
-    [SerializeField] private TMP_Text scoreText;
-
-
-    public void SetYakuInfo(string yakuName, string score)
-    {
-        yakuNameText.text = yakuName;
-        scoreText.text = score;
-    }
-}
